changed variable type to integer

+++ project.cpp
@@ -6,9 +6,9 @@ void integer_addition() //function whose execution time is to be measured
 {
 	long long loop_max = 10000000000;
   long long loop_no = 0;
-  long a = 2;
-  long b = 4;
-  long sum;
+  int a = 2;
+  int b = 4;
+  int sum;
 
   while (loop_no < loop_max)
   {
